Session validations added
v1.2.2

+++ src/Controllers/PublicProfile/Backend/FeedbackController.php
@@ -12,7 +12,7 @@ use So2platform\Publicprofile\Models\Post;
 use So2platform\Publicprofile\Models\ProfileFeedback;
 use So2platform\Publicprofile\Models\PublicProfile;
 
-class FeedbackController extends Controller
+class FeedbackController extends BackendBaseController
 {
     /**
      * Show the application index.
@@ -21,30 +21,20 @@ class FeedbackController extends Controller
      */
     public function index()
     {
-        /* If there's a active session then get the id from it. */
-        if(!empty(auth(config('publicprofile.auth_guard'))->user())){
-            // Change session id name.
-            $user_id = auth(config('publicprofile.auth_guard'))->user()[config('publicprofile.auth_model_key')];
-        }else{
-            /* Use your own logic to set the user id to the new post */
-            $user_id = config('publicprofile.default_auth_model_id'); // test id.
-        }
-        /* Get public profile with the session user_id */
-        $public_profile = PublicProfile::where('user_id', $user_id)->first();
+        $public_profile = PublicProfile::where('user_id', $this->auth_user_id )->first();
         if(!empty($public_profile)) {
             $profile_feedback = ProfileFeedback::where('public_profile_id', $public_profile->id)->first();
         }
         if(!empty($profile_feedback)){
             $feedbacks = Feedback::where('profile_feedback_id', $profile_feedback->id)->get();
-        }else {
+            if( $feedbacks->isEmpty() ){
+                return view("publicprofile::backend.layout.error", array(
+                    'error' => $this->no_feedbacks
+                ));
+            }
+        } else {
             return view("publicprofile::backend.layout.error", array(
-                'error' => "
-                        <div>
-                            <h1>Sin perfil</h1>
-                            <a class='btn btn-default' href='".route('backend_profile_create')."'>
-                                Crear
-                            </a>
-                        </div>"
+                'error' => $this->no_feedbacks
             ));
         }
         return view('publicprofile::backend.feedback.index', array(
@@ -59,11 +49,17 @@ class FeedbackController extends Controller
      */
     public function update($feedback_id)
     {
-        $feedback = Feedback::find($feedback_id);
-        if($feedback != null){
-            $feedback->status = $feedback->status == true ? false : true;
-            $feedback->save();
-            event(new FeedbacksEvent('feedback_'.$feedback->profile_feedback_id));
+        $public_profile = PublicProfile::where('user_id', $this->auth_user_id )->first();
+        if(!empty($public_profile)) {
+            $profile_feedback = ProfileFeedback::where('public_profile_id', $public_profile->id)->first();
+            if(!empty($profile_feedback)){
+                $feedback = Feedback::where('profile_feedback_id', $profile_feedback->id)->where('id',$feedback_id)->first();
+                if(!empty($feedback)){
+                    $feedback->status = $feedback->status == true ? false : true;
+                    $feedback->save();
+                    event(new FeedbacksEvent('feedback_'. $feedback->profile_feedback_id));
+                }
+            }
         }
         return back();
     }
@@ -75,14 +71,18 @@ class FeedbackController extends Controller
      */
     public function destroy($feedback_id)
     {
-        $feedback = Feedback::find($feedback_id);
-        if($feedback != null){
-            event(new FeedbacksEvent('feedback_'.$feedback->profile_feedback_id));
+        $public_profile = PublicProfile::where('user_id', $this->auth_user_id )->first();
+        if(!empty($public_profile)) {
+            $profile_feedback = ProfileFeedback::where('public_profile_id', $public_profile->id)->first();
+            if(!empty($profile_feedback)){
+                $feedback = Feedback::where('profile_feedback_id', $profile_feedback->id)->where('id',$feedback_id)->first();
+                if(!empty($feedback)){
+                    event(new FeedbacksEvent('feedback_'.$feedback->profile_feedback_id));
+                    $feedback->delete();
+                }
+            }
         }
-        $feedback->delete();
         return back();
     }
-
-
 
 }
